+++ dsp/phase_distortion/default.cpp
@@ -2,6 +2,6 @@
 
 void DefaultAlgo::update_params_() {}
 
-uint32_t DefaultAlgo::process_phase(uint32_t pha) {
+float DefaultAlgo::process_phase(float pha) {
     return pha;
 }
